+++ processor/src/test/fixtures/expected/com/example/prop/React4j_MutableDisposableActAsComponentTrackingProp.java
@@ -110,7 +110,7 @@ abstract class React4j_MutableDisposableActAsComponentTrackingProp extends Mutab
     $$react4j$$_state = ViewState.IDLE;
     assert Disposable.isNotDisposed( this );
     final MutableDisposableActAsComponentTrackingProp.ActAsComponentComponent $$react4jv$$_model = model();
-    if ( null != $$react4jv$$_model && !ComponentObservable.observe( $$react4jv$$_model ) ) {
+    if ( $$react4jv$$_model instanceof ComponentObservable && !ComponentObservable.observe( $$react4jv$$_model ) ) {
       return null;
     }
     if ( Disposable.isDisposed( $$react4jv$$_model ) ) {