+++ processor/src/test/fixtures/expected/com/example/prop/React4j_ObserveOnRenderComponentObservableNonTrackingProp.java
@@ -1,0 +1,112 @@
+package com.example.prop;
+
+import arez.Disposable;
+import arez.annotations.ArezComponent;
+import arez.annotations.Feature;
+import arez.annotations.SuppressArezWarnings;
+import java.util.Objects;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import javax.annotation.processing.Generated;
+import jsinterop.annotations.JsConstructor;
+import jsinterop.base.Js;
+import jsinterop.base.JsPropertyMap;
+import react4j.React;
+import react4j.ReactNode;
+import react4j.internal.OnComponentWillUnmount;
+import react4j.internal.OnShouldComponentUpdate;
+import react4j.internal.ViewConstructorFunction;
+
+@SuppressArezWarnings("Arez:UnnecessaryAllowEmpty")
+@ArezComponent(
+    name = "com_example_prop_ObserveOnRenderComponentObservableNonTrackingProp",
+    disposeNotifier = Feature.DISABLE,
+    sting = Feature.DISABLE,
+    allowEmpty = true
+)
+@Generated("react4j.processor.React4jProcessor")
+abstract class React4j_ObserveOnRenderComponentObservableNonTrackingProp extends ObserveOnRenderComponentObservableNonTrackingProp {
+  @Nonnull
+  private final react4j.internal.NativeView $$react4j$$_nativeView;
+
+  React4j_ObserveOnRenderComponentObservableNonTrackingProp(
+      @Nonnull final react4j.internal.NativeView $$react4j$$_nativeView) {
+    this.$$react4j$$_nativeView = Objects.requireNonNull( $$react4j$$_nativeView );
+  }
+
+  @Nonnull
+  private static ViewConstructorFunction getConstructorFunction() {
+    final ViewConstructorFunction viewConstructor = NativeView::new;
+    if ( React.enableViewNames() ) {
+      Js.asPropertyMap( viewConstructor ).set( "displayName", "ObserveOnRenderComponentObservableNonTrackingProp" );
+    }
+    return viewConstructor;
+  }
+
+  @Nullable
+  @Override
+  ObserveOnRenderComponentObservableNonTrackingProp.Model getModel() {
+    if ( React.shouldCheckInvariants() ) {
+      return null != $$react4j$$_nativeView.inputs().getAsAny( Inputs.model ) ? $$react4j$$_nativeView.inputs().getAsAny( Inputs.model ).cast() : null;
+    } else {
+      return Js.uncheckedCast( $$react4j$$_nativeView.inputs().getAsAny( Inputs.model ) );
+    }
+  }
+
+  private boolean $$react4j$$_shouldComponentUpdate(
+      @Nullable final JsPropertyMap<Object> nextInputs) {
+    assert null != nextInputs;
+    final JsPropertyMap<Object> inputs = $$react4j$$_nativeView.inputs();
+    if ( !Js.isTripleEqual( inputs.get( Inputs.model ), nextInputs.get( Inputs.model ) ) ) {
+      return true;
+    }
+    return false;
+  }
+
+  private void $$react4j$$_componentWillUnmount() {
+    ((Arez_React4j_ObserveOnRenderComponentObservableNonTrackingProp) this).dispose();
+  }
+
+  @Nullable
+  ReactNode $$react4j$$_render() {
+    assert Disposable.isNotDisposed( this );
+    final ObserveOnRenderComponentObservableNonTrackingProp.Model $$react4jv$$_getModel = getModel();
+    return render();
+  }
+
+  static final class Factory {
+    @Nonnull
+    static final ViewConstructorFunction TYPE = getConstructorFunction();
+  }
+
+  static final class Inputs {
+    static final String model = React.shouldMinimizeInputKeys() ? "a" : "model";
+  }
+
+  private static final class NativeView extends react4j.internal.NativeView implements OnShouldComponentUpdate, OnComponentWillUnmount {
+    @Nonnull
+    private final React4j_ObserveOnRenderComponentObservableNonTrackingProp $$react4j$$_view;
+
+    @JsConstructor
+    NativeView(@Nullable final JsPropertyMap<Object> inputs) {
+      super( inputs );
+      $$react4j$$_view = new Arez_React4j_ObserveOnRenderComponentObservableNonTrackingProp( this );
+    }
+
+    @Override
+    public final boolean shouldComponentUpdate(@Nonnull final JsPropertyMap<Object> nextInputs) {
+      return $$react4j$$_view.$$react4j$$_shouldComponentUpdate( nextInputs );
+    }
+
+    @Override
+    public final void componentWillUnmount() {
+      $$react4j$$_view.$$react4j$$_componentWillUnmount();
+    }
+
+    @Override
+    @Nullable
+    public final ReactNode render() {
+      return $$react4j$$_view.$$react4j$$_render();
+    }
+  }
+}