+++ processor/src/test/fixtures/input/com/example/prop/MutableDisposableArezComponentMaybeTrackingProp.java
@@ -11,7 +11,7 @@ import react4j.annotations.View;
 @View( type = View.Type.MAYBE_TRACKING )
 abstract class MutableDisposableArezComponentMaybeTrackingProp
 {
-  @ArezComponent( allowEmpty = true )
+  @ArezComponent( allowEmpty = true, disposeOnDeactivate = true )
   static abstract class Model
   {
   }