+++ processor/src/test/fixtures/input/com/example/prop/ObserveOnRenderAutodetectInterfaceTrackingProp.java
@@ -1,23 +1,21 @@
 package com.example.prop;
 
-import arez.annotations.ArezComponent;
-import arez.annotations.Feature;
 import javax.annotation.Nullable;
 import react4j.ReactNode;
+import react4j.annotations.Feature;
 import react4j.annotations.Input;
 import react4j.annotations.Render;
 import react4j.annotations.View;
 
 @View( type = View.Type.TRACKING )
-abstract class MutableDisposableNullableArezComponentTrackingProp
+abstract class ObserveOnRenderAutodetectInterfaceTrackingProp
 {
-  @ArezComponent( allowEmpty = true, observable = Feature.ENABLE )
-  static abstract class Model
+  interface Model
   {
   }
 
   @Nullable
-  @Input
+  @Input( disposable = Feature.ENABLE )
   abstract Model getModel();
 
   @Nullable