+++ processor/src/test/fixtures/input/com/example/prop/ObserveOnRenderComponentObservableNonTrackingProp.java
@@ -1,18 +1,17 @@
 package com.example.prop;
 
-import arez.annotations.ArezComponent;
-import arez.annotations.Feature;
+import arez.component.ComponentObservable;
 import javax.annotation.Nullable;
 import react4j.ReactNode;
 import react4j.annotations.Input;
 import react4j.annotations.Render;
 import react4j.annotations.View;
 
-@View( type = View.Type.TRACKING )
-abstract class MutableDisposableNullableArezComponentTrackingProp
+@View
+abstract class ObserveOnRenderComponentObservableNonTrackingProp
 {
-  @ArezComponent( allowEmpty = true, observable = Feature.ENABLE )
-  static abstract class Model
+  interface Model
+    extends ComponentObservable
   {
   }