+++ processor/src/test/fixtures/input/com/example/prop/ObserveOnRenderComponentObservableTrackingProp.java
@@ -1,7 +1,6 @@
 package com.example.prop;
 
-import arez.annotations.ArezComponent;
-import arez.annotations.Feature;
+import arez.component.ComponentObservable;
 import javax.annotation.Nullable;
 import react4j.ReactNode;
 import react4j.annotations.Input;
@@ -9,10 +8,10 @@ import react4j.annotations.Render;
 import react4j.annotations.View;
 
 @View( type = View.Type.TRACKING )
-abstract class MutableDisposableNullableArezComponentTrackingProp
+abstract class ObserveOnRenderComponentObservableTrackingProp
 {
-  @ArezComponent( allowEmpty = true, observable = Feature.ENABLE )
-  static abstract class Model
+  interface Model
+    extends ComponentObservable
   {
   }